fix: moved userUtils to utils
